refactor: WireTransaction decoding
