+++ plasma_framework/contracts/mocks/transactions/PaymentTransactionModelMock.sol
@@ -4,10 +4,23 @@ pragma experimental ABIEncoderV2;
 import "../../src/transactions/PaymentTransactionModel.sol";
 
 contract PaymentTransactionModelMock {
+    using RLPReader for bytes;
 
-    function decode(bytes memory _transaction) public pure returns (PaymentTransactionModel.Transaction memory) {
-        PaymentTransactionModel.Transaction memory transaction = PaymentTransactionModel.decode(_transaction);
-        return transaction;
+    function decode(bytes memory transaction) public pure returns (PaymentTransactionModel.Transaction memory) {
+        return PaymentTransactionModel.decode(transaction);
     }
 
+    function decodeOutput(bytes memory output) public pure returns (WireTransaction.Output memory) {
+        return PaymentTransactionModel.decodeOutput(output.toRlpItem());
+    }
+
+    function getOutputOwner(uint256 outputType, address owner, address token, uint256 amount) public pure returns (address payable) {
+        WireTransaction.Output memory output = WireTransaction.Output({
+            outputType: outputType,
+            outputGuard: bytes20(owner),
+            token: token,
+            amount: amount
+        });
+        return PaymentTransactionModel.getOutputOwner(output);
+    }
 }
