+++ plasma_framework/contracts/src/exits/payment/controllers/PaymentChallengeIFENotCanonical.sol
@@ -102,7 +102,10 @@ library PaymentChallengeIFENotCanonical {
         require(outputId == ife.inputs[args.inFlightTxInputIndex].outputId,
                 "Provided inputs data does not point to the same outputId from the in-flight exit");
 
-        WireTransaction.Output memory output = WireTransaction.getOutput(args.inputTx, inputUtxoPos.outputIndex());
+        WireTransaction.Output memory output = WireTransaction.getOutput(
+            WireTransaction.decode(args.inputTx),
+            inputUtxoPos.outputIndex()
+        );
 
         ISpendingCondition condition = self.spendingConditionRegistry.spendingConditions(
             output.outputType, self.supportedTxType
