+++ plasma_framework/contracts/src/exits/payment/controllers/PaymentChallengeStandardExit.sol
@@ -16,7 +16,6 @@ import "../../../utils/SafeEthTransfer.sol";
 import "../../../utils/UtxoPosLib.sol";
 import "../../../transactions/PaymentTransactionModel.sol";
 import "../../../transactions/WireTransaction.sol";
-import "../../../transactions/outputs/PaymentOutputModel.sol";
 
 library PaymentChallengeStandardExit {
     using UtxoPosLib for UtxoPosLib.UtxoPos;
@@ -40,6 +39,7 @@ library PaymentChallengeStandardExit {
         Controller controller;
         PaymentStandardExitRouterArgs.ChallengeStandardExitArgs args;
         PaymentExitDataModel.StandardExit exitData;
+        uint256 challengeTxType;
     }
 
     /**
@@ -77,10 +77,13 @@ library PaymentChallengeStandardExit {
     )
         public
     {
+        WireTransaction.Transaction memory challengeTx = WireTransaction.decode(args.challengeTx);
+
         ChallengeStandardExitData memory data = ChallengeStandardExitData({
             controller: self,
             args: args,
-            exitData: exitMap.exits[args.exitId]
+            exitData: exitMap.exits[args.exitId],
+            challengeTxType: challengeTx.txType
         });
         verifyChallengeExitExists(data);
         verifyChallengeTxProtocolFinalized(data);
@@ -108,13 +111,12 @@ library PaymentChallengeStandardExit {
         PaymentStandardExitRouterArgs.ChallengeStandardExitArgs memory args = data.args;
 
         UtxoPosLib.UtxoPos memory utxoPos = UtxoPosLib.UtxoPos(data.exitData.utxoPos);
-        PaymentOutputModel.Output memory output = PaymentTransactionModel
+        WireTransaction.Output memory output = PaymentTransactionModel
             .decode(args.exitingTx)
             .outputs[utxoPos.outputIndex()];
 
-        uint256 challengeTxType = WireTransaction.getTransactionType(args.challengeTx);
         ISpendingCondition condition = data.controller.spendingConditionRegistry.spendingConditions(
-            output.outputType, challengeTxType
+            output.outputType, data.challengeTxType
         );
         require(address(condition) != address(0), "Spending condition contract not found");
