+++ plasma_framework/contracts/src/exits/utils/MoreVpFinalization.sol
@@ -28,10 +28,10 @@ library MoreVpFinalization {
         view
         returns (bool)
     {
-        uint256 txType = WireTransaction.getTransactionType(txBytes);
-        uint8 protocol = framework.protocols(txType);
+        WireTransaction.Transaction memory wtx = WireTransaction.decode(txBytes);
+        uint8 protocol = framework.protocols(wtx.txType);
         require(protocol == Protocol.MORE_VP(), "MoreVpFinalization: not a MoreVP protocol tx");
-        
+
         (bytes32 root,) = framework.blocks(txPos.blockNum());
         require(root != bytes32(""), "Failed to get the root hash of the block num");
 
@@ -54,14 +54,14 @@ library MoreVpFinalization {
     )
         internal
         view
-        returns (bool) 
+        returns (bool)
     {
         if (txBytes.length == 0) {
             return false;
         }
 
-        uint256 txType = WireTransaction.getTransactionType(txBytes);
-        uint8 protocol = framework.protocols(txType);
+        WireTransaction.Transaction memory wtx = WireTransaction.decode(txBytes);
+        uint8 protocol = framework.protocols(wtx.txType);
         require(protocol == Protocol.MORE_VP(), "MoreVpFinalization: not a MoreVP protocol tx");
 
         return true;
