+++ plasma_framework/contracts/src/transactions/PaymentTransactionModel.sol
@@ -1,26 +1,24 @@
 pragma solidity 0.5.11;
+pragma experimental ABIEncoderV2;
 
-import "./outputs/PaymentOutputModel.sol";
+import "./WireTransaction.sol";
+import "../utils/AddressPayable.sol";
 import "../utils/RLPReader.sol";
 
 /**
  * @notice Data structure and its decode function for Payment transaction
  */
 library PaymentTransactionModel {
-
     using RLPReader for bytes;
     using RLPReader for RLPReader.RLPItem;
-    using PaymentOutputModel for PaymentOutputModel.Output;
 
     uint8 constant public MAX_INPUT_NUM = 4;
     uint8 constant public MAX_OUTPUT_NUM = 4;
 
-    uint8 constant private ENCODED_LENGTH = 4;
-
     struct Transaction {
         uint256 txType;
         bytes32[] inputs;
-        PaymentOutputModel.Output[] outputs;
+        WireTransaction.Output[] outputs;
         bytes32 metaData;
     }
 
@@ -41,36 +39,47 @@ library PaymentTransactionModel {
      * @return A decoded PaymentTransaction struct
      */
     function decode(bytes memory _tx) internal pure returns (PaymentTransactionModel.Transaction memory) {
-        RLPReader.RLPItem[] memory rlpTx = _tx.toRlpItem().toList();
-        require(rlpTx.length == ENCODED_LENGTH, "Invalid encoding of transaction");
+        WireTransaction.Transaction memory btx = WireTransaction.decode(_tx);
+        return decodeFromWire(btx);
+    }
 
-        RLPReader.RLPItem[] memory rlpInputs = rlpTx[1].toList();
-        require(rlpInputs.length <= MAX_INPUT_NUM, "Transaction inputs num exceeds limit");
+    function decodeFromWire(WireTransaction.Transaction memory btx) internal pure returns (PaymentTransactionModel.Transaction memory) {
+        require(btx.inputs.length <= MAX_INPUT_NUM, "Transaction inputs num exceeds limit");
+        require(btx.outputs.length <= MAX_OUTPUT_NUM, "Transaction outputs num exceeds limit");
 
-        RLPReader.RLPItem[] memory rlpOutputs = rlpTx[2].toList();
-        require(rlpOutputs.length <= MAX_OUTPUT_NUM, "Transaction outputs num exceeds limit");
-
-        uint txType = rlpTx[0].toUint();
-        require(txType > 0, "Transaction type must not be 0");
-
-        bytes32[] memory inputs = new bytes32[](rlpInputs.length);
-        for (uint i = 0; i < rlpInputs.length; i++) {
-            bytes32 input = rlpInputs[i].toBytes32();
+        bytes32[] memory inputs = new bytes32[](btx.inputs.length);
+        for (uint i = 0; i < btx.inputs.length; i++) {
+            bytes32 input = btx.inputs[i].toBytes32();
             // Disallow null inputs
             require(uint256(input) != 0, "Null input not allowed");
             inputs[i] = input;
         }
 
-        PaymentOutputModel.Output[] memory outputs = new PaymentOutputModel.Output[](rlpOutputs.length);
-        for (uint i = 0; i < rlpOutputs.length; i++) {
-            PaymentOutputModel.Output memory output = PaymentOutputModel.decode(rlpOutputs[i]);
-            require(output.outputType != 0, "Output type must not be 0");
-            require(output.amount != 0, "Output amount must not be 0");
-            outputs[i] = output;
+        WireTransaction.Output[] memory outputs = new WireTransaction.Output[](btx.outputs.length);
+        for (uint i = 0; i < btx.outputs.length; i++) {
+            outputs[i] = decodeOutput(btx.outputs[i]);
         }
 
-        bytes32 metaData = rlpTx[3].toBytes32();
+        bytes32 metaData = btx.txData.toBytes32();
 
-        return Transaction({txType: txType, inputs: inputs, outputs: outputs, metaData: metaData});
+        return Transaction({txType: btx.txType, inputs: inputs, outputs: outputs, metaData: metaData});
+    }
+
+    function decodeOutput(RLPReader.RLPItem memory output) internal pure returns (WireTransaction.Output memory) {
+        RLPReader.RLPItem[] memory outputRlpList = output.toList();
+        require(outputRlpList.length == 4, "Output must have 4 items");
+        WireTransaction.Output memory decodedOutput = WireTransaction.decodeOutput(outputRlpList);
+        require(decodedOutput.amount != 0, "Output amount must not be 0");
+        return decodedOutput;
+    }
+
+    /**
+     * @notice Retrieve the 'owner' from the output, assuming the
+     *         'outputGuard' field directly holds the owner's address
+     * @dev It's possible that 'outputGuard' can be a hash of preimage that holds the owner information,
+     *       but this should not and cannot be handled here.
+     */
+    function getOutputOwner(WireTransaction.Output memory _output) internal pure returns (address payable) {
+        return AddressPayable.convert(address(_output.outputGuard));
     }
 }
