+++ plasma_framework/contracts/src/vaults/verifiers/EthDepositVerifier.sol
@@ -2,14 +2,11 @@ pragma solidity 0.5.11;
 
 import "./IEthDepositVerifier.sol";
 import {PaymentTransactionModel as DepositTx} from "../../transactions/PaymentTransactionModel.sol";
-import {PaymentOutputModel as DepositOutputModel} from "../../transactions/outputs/PaymentOutputModel.sol";
 
 /**
  * @notice Implementation of ETH deposit verifier using payment transaction as the deposit transaction
  */
 contract EthDepositVerifier is IEthDepositVerifier {
-    using DepositOutputModel for DepositOutputModel.Output;
-
     uint256 public depositTxType;
     uint256 public supportedOutputType;
 
@@ -34,7 +31,8 @@ contract EthDepositVerifier is IEthDepositVerifier {
         require(decodedTx.outputs[0].token == address(0), "Output requires correct currency (ETH)");
         require(decodedTx.outputs[0].outputType == supportedOutputType, "Invalid output type");
 
-        address depositorsAddress = decodedTx.outputs[0].owner();
+        // address depositorsAddress = decodedTx.outputs[0].owner();
+        address depositorsAddress = DepositTx.getOutputOwner(decodedTx.outputs[0]);
         require(depositorsAddress == sender, "Depositor's address must match sender's address");
     }
 }
